bug fix: now successfully handles song sets
There was a condition statement that checked to see if the page
only contained one song, but it reported true if the page was a
set of songs. This made sc only download or list the first song
in a set. The conditional now checks to make sure the page is not
a set before reporting true.